backend/requirements.txt
fastapi
uvicorn
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
python-multipart
httpx
openai
pydantic
python-dotenv
beautifulsoup4
